+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 * [#488] Fixes for locked doors and missing monsters in The Lost Citadel of the Scarlet Minotaur adventure
 
 ### Enhancement
+* [#342] Identify "rare" languages in the language selector to aid in selection
 * [#447] Added ability to easily access filtered compendium items
 * [#480] Added ability to create custom weapon and armor properties
 * [#490] Added Condition for Halfling's "Stealthy" Invisibility ability